+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/abc/types/traits/Trait.java
@@ -366,7 +366,9 @@ public abstract class Trait implements Cloneable, Serializable {
         writer.hilightSpecial(traitType, HighlightSpecialType.TRAIT_TYPE);
         writer.appendNoHilight(" ");
         writer.hilightSpecial(abc.constants.multinameToString(name_index), HighlightSpecialType.TRAIT_NAME);
-        writer.indent();
+        if (Configuration.indentAs3PCode.get()) {
+            writer.indent();
+        }
         if ((kindFlags & ATTR_Final) > 0) {
             writer.newLine();
             writer.append("flag ");
@@ -394,7 +396,9 @@ public abstract class Trait implements Cloneable, Serializable {
                 writer.append("\"");
                 writer.append(Helper.escapeActionScriptString(abc.constants.getString(abc.metadata_info.get(m).name_index)));
                 writer.append("\"");
-                writer.indent();
+                if (Configuration.indentAs3PCode.get()) {
+                    writer.indent();
+                }
                 writer.newLine();
                 if (m >= 0 && m < abc.metadata_info.size()) {
                     for (int i = 0; i < abc.metadata_info.get(m).keys.length; i++) {
@@ -414,7 +418,9 @@ public abstract class Trait implements Cloneable, Serializable {
                         writer.newLine();
                     }
                 }
-                writer.unindent();
+                if (Configuration.indentAs3PCode.get()) {
+                    writer.unindent();
+                }
                 writer.append("end ; metadata");
             }
         }