+++ src/com/jpexs/decompiler/flash/gui/abc/ASMSourceEditorPane.java
@@ -124,7 +124,10 @@ public class ASMSourceEditorPane extends DebuggableEditorPane implements CaretLi
         MethodBody body = abc.bodies.get(bodyIndex);
         abc.bodies.get(bodyIndex).getCode().toASMSource(abc, abc.constants, abc.method_info.get(body.method_info), body, exportMode, writer);
         if (trait != null && exportMode != ScriptExportMode.AS && exportMode != ScriptExportMode.AS_METHOD_STUBS) {
-            writer.unindent().appendNoHilight("end ; trait").newLine();
+            if (Configuration.indentAs3PCode.get()) {
+                writer.unindent();
+            }
+            writer.appendNoHilight("end ; trait").newLine();
         }
         return new HighlightedText(writer);
     }